+++ openstack/identity/v3/extensions/trusts/results.go
@@ -1,6 +1,11 @@
 package trusts
 
-import "github.com/gophercloud/gophercloud"
+import (
+	"time"
+
+	"github.com/gophercloud/gophercloud"
+	"github.com/gophercloud/gophercloud/pagination"
+)
 
 type trustResult struct {
 	gophercloud.Result
@@ -18,34 +23,72 @@ type DeleteResult struct {
 	gophercloud.ErrResult
 }
 
+// TrustPage is a single page of Region results.
+type TrustPage struct {
+	pagination.LinkedPageBase
+}
+
+// GetResult is the response from a Get operation. Call its Extract method
+// to interpret it as a Trust.
+type GetResult struct {
+	trustResult
+}
+
+// IsEmpty determines whether or not a page of Trusts contains any results.
+func (t TrustPage) IsEmpty() (bool, error) {
+	roles, err := ExtractTrusts(t)
+	return len(roles) == 0, err
+}
+
+// NextPageURL extracts the "next" link from the links section of the result.
+func (t TrustPage) NextPageURL() (string, error) {
+	var s struct {
+		Links struct {
+			Next     string `json:"next"`
+			Previous string `json:"previous"`
+		} `json:"links"`
+	}
+	err := t.ExtractInto(&s)
+	if err != nil {
+		return "", err
+	}
+	return s.Links.Next, err
+}
+
+// ExtractProjects returns a slice of Trusts contained in a single page of
+// results.
+func ExtractTrusts(r pagination.Page) ([]Trust, error) {
+	var s struct {
+		Trusts []Trust `json:"trusts"`
+	}
+	err := (r.(TrustPage)).ExtractInto(&s)
+	return s.Trusts, err
+}
+
 // Extract interprets any trust result as a Trust.
-func (r trustResult) Extract() (*Trust, error) {
+func (t trustResult) Extract() (*Trust, error) {
 	var s struct {
 		Trust *Trust `json:"trust"`
 	}
-	err := r.ExtractInto(&s)
+	err := t.ExtractInto(&s)
 	return s.Trust, err
-}
-
-// TrusteeUser represents the trusted user ID of a trust.
-type TrusteeUser struct {
-	ID string `json:"id"`
-}
-
-// TrustorUser represents the trusting user ID of a trust.
-type TrustorUser struct {
-	ID string `json:"id"`
 }
 
 // Trust represents a delegated authorization request between two
 // identities.
 type Trust struct {
-	ID                 string      `json:"id"`
-	Impersonation      bool        `json:"impersonation"`
-	TrusteeUser        TrusteeUser `json:"trustee_user"`
-	TrustorUser        TrustorUser `json:"trustor_user"`
-	RedelegatedTrustID string      `json:"redelegated_trust_id"`
-	RedelegationCount  int         `json:"redelegation_count"`
+	ID                 string    `json:"id"`
+	Impersonation      bool      `json:"impersonation"`
+	TrusteeUserID      string    `json:"trustee_user_id"`
+	TrustorUserID      string    `json:"trustor_user_id"`
+	RedelegatedTrustID string    `json:"redelegated_trust_id"`
+	RedelegationCount  int       `json:"redelegation_count,omitempty"`
+	AllowRedelegation  bool      `json:"allow_redelegation,omitempty"`
+	ProjectID          string    `json:"project_id,omitempty"`
+	RemainingUses      bool      `json:"remaining_uses,omitempty"`
+	Roles              []Role    `json:"roles,omitempty"`
+	DeletedAt          time.Time `json:"deleted_at"`
+	ExpiresAt          time.Time `json:"expires_at"`
 }
 
 // Role specifies a single role that is granted to a trustee.