+++ openstack/identity/v3/extensions/trusts/testing/fixtures.go
@@ -4,7 +4,9 @@ import (
 	"fmt"
 	"net/http"
 	"testing"
+	"time"
 
+	"github.com/gophercloud/gophercloud/openstack/identity/v3/extensions/trusts"
 	"github.com/gophercloud/gophercloud/openstack/identity/v3/tokens"
 	"github.com/gophercloud/gophercloud/testhelper"
 	"github.com/gophercloud/gophercloud/testhelper/client"
@@ -49,6 +51,66 @@ const CreateResponse = `
         "trustee_user_id": "ecb37e88cc86431c99d0332208cb6fbf",
         "trustor_user_id": "959ed913a32c4ec88c041c98e61cbbc3"
     }
+}
+`
+
+// GetOutput provides a Get result.
+const GetResponse = `
+{
+    "trust": {
+        "id": "987fe8",
+        "expires_at": "2013-02-27T18:30:59.999999Z",
+        "impersonation": true,
+        "links": {
+            "self": "http://example.com/identity/v3/OS-TRUST/trusts/987fe8"
+        },
+        "roles": [
+            {
+                "id": "ed7b78",
+                "links": {
+                    "self": "http://example.com/identity/v3/roles/ed7b78"
+                },
+                "name": "member"
+            }
+        ],
+        "roles_links": {
+            "next": null,
+            "previous": null,
+            "self": "http://example.com/identity/v3/OS-TRUST/trusts/1ff900/roles"
+        },
+        "project_id": "0f1233",
+        "trustee_user_id": "be34d1",
+        "trustor_user_id": "56ae32"
+    }
+}
+`
+
+// ListOutput provides a single page of Role results.
+const ListResponse = `
+{
+    "trusts": [
+        {
+            "id": "1ff900",
+            "expires_at": "2013-02-27T18:30:59.999999Z",
+            "impersonation": true,
+            "links": {
+                "self": "http://example.com/identity/v3/OS-TRUST/trusts/1ff900"
+            },
+            "project_id": "0f1233",
+            "trustee_user_id": "86c0d5",
+            "trustor_user_id": "a0fdfd"
+        },
+        {
+            "id": "f4513a",
+            "impersonation": false,
+            "links": {
+                "self": "http://example.com/identity/v3/OS-TRUST/trusts/f45513a"
+            },
+            "project_id": "0f1233",
+            "trustee_user_id": "86c0d5",
+            "trustor_user_id": "3cd2ce"
+        }
+    ]
 }
 `
 
@@ -131,5 +193,56 @@ func HandleDeleteTrust(t *testing.T) {
 		testhelper.TestHeader(t, r, "X-Auth-Token", client.TokenID)
 
 		w.WriteHeader(http.StatusNoContent)
+	})
+}
+
+// HandleGetTrustSuccessfully creates an HTTP handler at `/OS-TRUST/trusts` on the
+// test handler mux that responds with a single trusts.
+func HandleGetTrustSuccessfully(t *testing.T) {
+	testhelper.Mux.HandleFunc("/OS-TRUST/trusts/987fe8", func(w http.ResponseWriter, r *http.Request) {
+		testhelper.TestMethod(t, r, "GET")
+		testhelper.TestHeader(t, r, "Accept", "application/json")
+		testhelper.TestHeader(t, r, "X-Auth-Token", client.TokenID)
+
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprintf(w, GetResponse)
+	})
+}
+
+var FirstTrust = trusts.Trust{
+	ID:            "1ff900",
+	Impersonation: true,
+	TrusteeUserID: "86c0d5",
+	TrustorUserID: "a0fdfd",
+	ProjectID:     "0f1233",
+	ExpiresAt:     time.Date(2013, 02, 27, 18, 30, 59, 999999000, time.UTC),
+	DeletedAt:     time.Time{},
+}
+
+var SecondTrust = trusts.Trust{
+	ID:            "f4513a",
+	Impersonation: false,
+	TrusteeUserID: "86c0d5",
+	TrustorUserID: "3cd2ce",
+	ProjectID:     "0f1233",
+	ExpiresAt:     time.Time{},
+	DeletedAt:     time.Time{},
+}
+
+// ExpectedRolesSlice is the slice of roles expected to be returned from ListOutput.
+var ExpectedTrustsSlice = []trusts.Trust{FirstTrust, SecondTrust}
+
+// HandleListTrustsSuccessfully creates an HTTP handler at `/OS-TRUST/trusts` on the
+// test handler mux that responds with a list of two trusts.
+func HandleListTrustsSuccessfully(t *testing.T) {
+	testhelper.Mux.HandleFunc("/OS-TRUST/trusts", func(w http.ResponseWriter, r *http.Request) {
+		testhelper.TestMethod(t, r, "GET")
+		testhelper.TestHeader(t, r, "Accept", "application/json")
+		testhelper.TestHeader(t, r, "X-Auth-Token", client.TokenID)
+
+		w.Header().Set("Content-Type", "application/json")
+		w.WriteHeader(http.StatusOK)
+		fmt.Fprintf(w, ListResponse)
 	})
 }