+++ openstack/identity/v3/extensions/trusts/testing/requests_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/gophercloud/gophercloud/openstack/identity/v3/extensions/trusts"
 	"github.com/gophercloud/gophercloud/openstack/identity/v3/tokens"
+	"github.com/gophercloud/gophercloud/pagination"
 	th "github.com/gophercloud/gophercloud/testhelper"
 	"github.com/gophercloud/gophercloud/testhelper/client"
 )
@@ -56,14 +57,8 @@ func TestCreateUserIDPasswordTrustID(t *testing.T) {
 		},
 		trusts.TokenExt{
 			Trust: trusts.Trust{
-				ID:            "fe0aef",
-				Impersonation: false,
-				TrusteeUser: trusts.TrusteeUser{
-					ID: "0ca8f6",
-				},
-				TrustorUser: trusts.TrustorUser{
-					ID: "bd263c",
-				},
+				ID:                 "fe0aef",
+				Impersonation:      false,
 				RedelegatedTrustID: "3ba234",
 				RedelegationCount:  2,
 			},
@@ -105,4 +100,59 @@ func TestDeleteTrust(t *testing.T) {
 
 	res := trusts.Delete(client.ServiceClient(), "3422b7c113894f5d90665e1a79655e23")
 	th.AssertNoErr(t, res.Err)
+}
+
+func TestGetTrust(t *testing.T) {
+	th.SetupHTTP()
+	defer th.TeardownHTTP()
+	HandleGetTrustSuccessfully(t)
+
+	res := trusts.Get(client.ServiceClient(), "987fe8")
+	th.AssertNoErr(t, res.Err)
+}
+
+func TestListTrusts(t *testing.T) {
+	th.SetupHTTP()
+	defer th.TeardownHTTP()
+	HandleListTrustsSuccessfully(t)
+
+	count := 0
+	err := trusts.List(client.ServiceClient(), nil).EachPage(func(page pagination.Page) (bool, error) {
+		count++
+
+		actual, err := trusts.ExtractTrusts(page)
+		th.AssertNoErr(t, err)
+
+		th.CheckDeepEquals(t, ExpectedTrustsSlice, actual)
+
+		return true, nil
+	})
+	th.AssertNoErr(t, err)
+	th.CheckEquals(t, count, 1)
+}
+
+func TestListTrustsAllPages(t *testing.T) {
+	th.SetupHTTP()
+	defer th.TeardownHTTP()
+	HandleListTrustsSuccessfully(t)
+
+	allPages, err := trusts.List(client.ServiceClient(), nil).AllPages()
+	th.AssertNoErr(t, err)
+	actual, err := trusts.ExtractTrusts(allPages)
+	th.AssertNoErr(t, err)
+	th.CheckDeepEquals(t, ExpectedTrustsSlice, actual)
+}
+
+func TestListTrustsFiltered(t *testing.T) {
+	th.SetupHTTP()
+	defer th.TeardownHTTP()
+	HandleListTrustsSuccessfully(t)
+	trustsListOpts := trusts.ListOpts{
+		TrustorUserID: "86c0d5",
+	}
+	allPages, err := trusts.List(client.ServiceClient(), trustsListOpts).AllPages()
+	th.AssertNoErr(t, err)
+	actual, err := trusts.ExtractTrusts(allPages)
+	th.AssertNoErr(t, err)
+	th.CheckDeepEquals(t, ExpectedTrustsSlice, actual)
 }