+++ openstack/placement/v1/resourceproviders/doc.go
@@ -22,7 +22,7 @@ Example to create resource providers
 	opts := resourceproviders.CreateOpts{
 		Name: "new-rp",
 		UUID: "b99b3ab4-3aa6-4fba-b827-69b88b9c544a",
-    }
+	}
 
 	rp, err := resourceproviders.Create(placementClient, opts).Extract()
 	if err != nil {